requirements.txt
Django==3.2.7
djangorestframework==3.12.4
python-decouple==3.4
